Refactor out city detail window
