+++ frontend/src/GamePage.js
@@ -27,6 +27,7 @@ import HexDisplay, { YieldImages } from './HexDisplay';
 import CityDisplay from './CityDisplay';
 import BuildingDisplay, { BriefBuildingDisplay, BriefBuildingDisplayTitle } from './BuildingDisplay';
 import UnitDisplay, {BriefUnitDisplay, BriefUnitDisplayTitle} from './UnitDisplay';
+import CityDetailWindow from './CityDetailWindow';
 import UpperRightDisplay from './UpperRightDisplay';
 import moveSound from './sounds/movement.mp3';
 import meleeAttackSound from './sounds/melee_attack.mp3';
@@ -206,33 +207,12 @@ function RulesDialog({open, onClose, gameConstants}) {
     )
 }
 
-export const FocusSelectorTitle = ({ title }) => {
-    return (
-        <div 
-            className="focus-title-card" 
-        >
-            <span>{title}</span>
-        </div>        
-    );
-}
-
 const generateUniqueId = () => {
     return Math.random().toString(36).substring(2);
 }
 
 const lowercaseAndReplaceSpacesWithUnderscores = (str) => {
     return str.toLowerCase().replace(/ /g, '_');
-}
-
-export const FocusSelectionOption = ({ focus, onClick, isSelected }) => {
-    return (
-        <div
-            className={`focus-selection-option ${focus} ${isSelected ? 'selected' : ''}`}
-            onClick={onClick}
-        >
-            <span>{focus}</span>
-        </div>
-    );
 }
 
 export default function GamePage() {
@@ -309,6 +289,7 @@ export default function GamePage() {
 
     const hoveredHexRef = React.useRef(hoveredHex);
     const lastSetPrimaryTargetRef = React.useRef(lastSetPrimaryTarget);
+    const playerApiUrl= `${URL}/api/player_input/${gameId}`
 
     useEffect(() => {
         hoveredHexRef.current = hoveredHex;
@@ -1846,118 +1827,7 @@ export default function GamePage() {
             });
     }
 
-    const handleClickBuildingChoice = (buildingName) => {
-        const playerInput = {
-            'building_name': (buildingName),
-            'move_type': 'choose_building',
-            'city_id': selectedCity.id,
-        }
-
-        const data = {
-            player_num: playerNum,
-            player_input: playerInput,
-        }
-
-        setHoveredBuilding(null);
-
-        fetch(`${URL}/api/player_input/${gameId}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(data),
-        }).then(response => response.json())
-            .then(data => {
-                if (data.game_state) {
-                    setGameState(data.game_state);
-                    refreshSelectedCity(data.game_state);
-                }
-            });
-    }
-
-    const handleCancelBuilding = (buildingName) => {
-        const playerInput = {
-            'building_name': (buildingName),
-            'move_type': 'cancel_building',
-            'city_id': selectedCity.id,
-        }
-
-        const data = {
-            player_num: playerNum,
-            player_input: playerInput,
-        }
-
-        setHoveredBuilding(null);
-
-        fetch(`${URL}/api/player_input/${gameId}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(data),
-        }).then(response => response.json())
-            .then(data => {
-                if (data.game_state) {
-                    setGameState(data.game_state);
-                    refreshSelectedCity(data.game_state);
-                }
-            });
-    }
-
-    const handleClickUnitChoice = (unitName) => {
-        const playerInput = {
-            'unit_name': (unitName),
-            'move_type': 'choose_unit',
-            'city_id': selectedCity.id,
-        }
-
-        const data = {
-            player_num: playerNum,
-            player_input: playerInput,
-        }
-
-        fetch(`${URL}/api/player_input/${gameId}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(data),
-        }).then(response => response.json())
-            .then(data => {
-                if (data.game_state) {
-                    setGameState(data.game_state);
-                    refreshSelectedCity(data.game_state);
-                }
-            });
-    }
-
-    const handleCancelUnit = (unitIndex) => {
-        const playerInput = {
-            'unit_index_in_queue': unitIndex,
-            'move_type': 'cancel_unit',
-            'city_id': selectedCity.id,
-        }
-
-        const data = {
-            player_num: playerNum,
-            player_input: playerInput,
-        }
-
-        fetch(`${URL}/api/player_input/${gameId}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(data),
-        }).then(response => response.json())
-            .then(data => {
-                if (data.game_state) {
-                    setGameState(data.game_state);
-                    refreshSelectedCity(data.game_state);
-                }
-            });
-    }
-
+    
     const handleClickTech = (tech) => {
 
         const playerInput = {
@@ -1969,7 +1839,7 @@ export default function GamePage() {
             player_num: playerNum,
             player_input: playerInput,
         }
-        fetch(`${URL}/api/player_input/${gameId}`, {
+        fetch(playerApiUrl, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
@@ -1983,33 +1853,7 @@ export default function GamePage() {
                 }
             });
     }
-
-    const handleClickFocus = (focus) => {
-        const playerInput = {
-            'city_id': selectedCity.id,
-            'focus': focus,
-            'move_type': 'choose_focus',
-        }
-
-        const data = {
-            player_num: playerNum,
-            player_input: playerInput,
-        }
-        fetch(`${URL}/api/player_input/${gameId}`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(data),
-        }).then(response => response.json())
-            .then(data => {
-                if (data.game_state) {
-                    setGameState(data.game_state);
-                    refreshSelectedCity(data.game_state);
-                }
-            });
-    }
-
+   
     useEffect(() => {
         fetch(`${URL}/api/civ_templates`)
             .then(response => response.json())
@@ -2204,7 +2048,6 @@ export default function GamePage() {
                     setTechChoices(data.tech_choices);
                     if (data.game_state) {
                         setGameState(data.game_state);
-                        // refreshSelectedCity(data.game_state);
                         refreshSelectedCity(data.game_state);
                     }
                 });
@@ -2216,7 +2059,7 @@ export default function GamePage() {
                     move_type: 'choose_decline_option',
                 },
             }
-            fetch(`${URL}/api/player_input/${gameId}`, {
+            fetch(playerApiUrl, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json'
@@ -2434,7 +2277,7 @@ export default function GamePage() {
                 player_input: playerInput,
             }
 
-            fetch(`${URL}/api/player_input/${gameId}`, {
+            fetch(playerApiUrl, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json'
@@ -2488,7 +2331,7 @@ export default function GamePage() {
             player_input: playerInput,
         }
 
-        fetch(`${URL}/api/player_input/${gameId}`, {
+        fetch(playerApiUrl, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
@@ -2516,7 +2359,7 @@ export default function GamePage() {
             player_input: playerInput,
         }
 
-        fetch(`${URL}/api/player_input/${gameId}`, {
+        fetch(playerApiUrl, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
@@ -2545,7 +2388,7 @@ export default function GamePage() {
             player_input: playerInput,
         }
 
-        fetch(`${URL}/api/player_input/${gameId}`, {
+        fetch(playerApiUrl, {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
@@ -2668,55 +2511,25 @@ export default function GamePage() {
                         nextForcedRollAt={gameState?.next_forced_roll_at}
                         gameId={gameId}
                     />}
-                    {selectedCityBuildingChoices && (
-                        <div className="building-choices-container">
-                            <BriefBuildingDisplayTitle title="Building Choices" />
-                            {selectedCityBuildingChoices.map((buildingName, index) => (
-                                <BriefBuildingDisplay key={index} buildingName={buildingName} unitTemplatesByBuildingName={unitTemplatesByBuildingName} buildingTemplates={buildingTemplates} setHoveredBuilding={setHoveredBuilding} onClick={() => handleClickBuildingChoice(buildingName)} descriptions={descriptions} />
-                            ))}
-                        </div>
-                    )};
-                    {selectedCityBuildingQueue && (
-                        <div className="building-queue-container">
-                            <BriefBuildingDisplayTitle title="Building Queue" />
-                            {selectedCityBuildingQueue.map((buildingName, index) => (
-                                <BriefBuildingDisplay key={index} buildingName={buildingName} unitTemplatesByBuildingName={unitTemplatesByBuildingName} buildingTemplates={buildingTemplates} setHoveredBuilding={setHoveredBuilding} onClick={() => handleCancelBuilding(buildingName)} descriptions={descriptions}/>
-                            ))}
-                        </div>
-                    )};                
-                    {selectedCityBuildings && (
-                        <div className="existing-buildings-container">
-                            <BriefBuildingDisplayTitle title="Existing buildings" />
-                            {selectedCityBuildings.map((buildingName, index) => (
-                                <BriefBuildingDisplay key={index} buildingName={buildingName} unitTemplatesByBuildingName={unitTemplatesByBuildingName} buildingTemplates={buildingTemplates} setHoveredBuilding={setHoveredBuilding} descriptions={descriptions}/>
-                            ))}
-                        </div>
-                    )};  
-                    {selectedCityUnitChoices && (
-                        <div className="unit-choices-container">
-                            <BriefUnitDisplayTitle title="Unit Choices" />
-                            {selectedCityUnitChoices.map((unitName, index) => (
-                                <BriefUnitDisplay key={index} unitName={unitName} unitTemplates={unitTemplates} setHoveredUnit={setHoveredUnit} onClick={() => handleClickUnitChoice(unitName)} />
-                            ))}
-                        </div>
-                    )};
-                    {selectedCityUnitQueue && (
-                        <div className="unit-queue-container">
-                            <BriefUnitDisplayTitle title="Unit Queue" />
-                            {selectedCityUnitQueue.map((unitName, index) => (
-                                <BriefUnitDisplay key={index} unitName={unitName} unitTemplates={unitTemplates} setHoveredUnit={setHoveredUnit} onClick={() => handleCancelUnit(index)}/>
-                            ))}
-                        </div>
-                    )};
-                    {selectedCity && !gameState?.special_mode_by_player_num?.[playerNum] && (
-                        <div className="focus-container">
-                            <FocusSelectorTitle title="City Focus" />
-                            <FocusSelectionOption focus="food" isSelected={selectedCity.focus === 'food'} onClick={() => handleClickFocus('food')} />
-                            <FocusSelectionOption focus="wood" isSelected={selectedCity.focus === 'wood'} onClick={() => handleClickFocus('wood')} />
-                            <FocusSelectionOption focus="metal" isSelected={selectedCity.focus === 'metal'} onClick={() => handleClickFocus('metal')} />
-                            <FocusSelectionOption focus="science" isSelected={selectedCity.focus === 'science'} onClick={() => handleClickFocus('science')} />
-                        </div>
-                    )}
+                    <CityDetailWindow 
+                        gameState={gameState}
+                        playerNum={playerNum}
+                        playerApiUrl={playerApiUrl}
+                        setGameState={setGameState}
+                        refreshSelectedCity={refreshSelectedCity}
+                        selectedCityBuildingChoices={selectedCityBuildingChoices} 
+                        selectedCityBuildingQueue={selectedCityBuildingQueue}
+                        selectedCityBuildings={selectedCityBuildings}
+                        selectedCityUnitChoices={selectedCityUnitChoices}
+                        selectedCityUnitQueue={selectedCityUnitQueue}
+                        selectedCity={selectedCity} 
+                        unitTemplatesByBuildingName={unitTemplatesByBuildingName}
+                        buildingTemplates={buildingTemplates}
+                        unitTemplates={unitTemplates}
+                        descriptions={descriptions}
+                        setHoveredUnit={setHoveredUnit}
+                        setHoveredBuilding={setHoveredBuilding}
+                        />
                     <div style={{position: 'fixed', top: '10px', left: '50%', transform: 'translate(-50%, 0%)'}}>                             
                         {hoveredBuilding && (
                             <BuildingDisplay buildingName={hoveredBuilding} unitTemplatesByBuildingName={unitTemplatesByBuildingName} buildingTemplates={buildingTemplates} />
